+++ domains/certificates/Query.API/API/Program.cs
@@ -7,7 +7,7 @@ using API.DataSyncSyncer;
 using API.GranularCertificateIssuer;
 using API.Query.API;
 using API.RabbitMq;
-using API.TransferCertificateService;
+using API.TransferCertificateHandler;
 using Marten;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
@@ -63,7 +63,7 @@ builder.Services.AddQueryApi();
 builder.Services.AddContractService();
 builder.Services.AddDataSyncSyncer(builder.Configuration);
 builder.Services.AddGranularCertificateIssuer();
-builder.Services.AddTransferCertificateService();
+builder.Services.AddTransferCertificateHandler();
 
 builder.Services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
     .AddJwtBearer(o =>
