+++ domains/certificates/Query.API/API/TransferCertificateHandler/Startup.cs
@@ -2,10 +2,10 @@ using API.GranularCertificateIssuer.Repositories;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
-namespace API.GranularCertificateIssuer;
+namespace API.TransferCertificateHandler;
 
 public static class Startup
 {
-    public static void AddGranularCertificateIssuer(this IServiceCollection services)
+    public static void AddTransferCertificateHandler(this IServiceCollection services)
         => services.TryAddSingleton<IProductionCertificateRepository, ProductionCertificateRepository>();
 }
