--- domains/certificates/Query.API/API/TransferCertificateService/Startup.cs
@@ -1,10 +1,0 @@
-using Microsoft.Extensions.DependencyInjection;
-
-namespace API.TransferCertificateService;
-
-public static class Startup
-{
-    public static void AddTransferCertificateService(this IServiceCollection services)
-    {
-    }
-}
